requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
bokeh==3.7.3
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorcet==3.1.0
contourpy==1.3.2
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.14
dotenv==0.9.9
fonttools==4.58.5
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
holoviews==1.21.0
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
hvplot==0.11.3
idna==3.10
ImageHash==4.3.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
linkify-it-py==2.0.3
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
multimethod==1.12
narwhals==1.45.0
networkx==3.5
numba==0.61.0
numpy==2.1.3
nvidia-nccl-cu12==2.27.6
packaging==24.2
pandas==2.3.0
panel==1.7.4
param==2.2.1
patsy==1.0.1
phik==0.12.4
pillow==11.3.0
protobuf==6.31.1
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyviz_comms==3.0.6
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.4
streamlit==1.45.1
streamlit-ydata-profiling==0.2.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
urllib3==2.5.0
visions==0.8.1
watchdog==3.0.0
webencodings==0.5.1
wordcloud==1.9.4
xgboost==3.0.2
xyzservices==2025.4.0
ydata-profiling==4.16.1
